requirements.txt
flask
flask-mysqldb
werkzeug
mysql-connector-python
python-dotenv
Pillow
cloudinary
msal
requests
gunicorn
